Flores Overhall
After working with artifacts from hobbits, I thought it would be nice
to let the user define groups. That said, whoa, this turned out to be
much tougher. I added the qualitative and quantitative inputs, but it
required a pretty thorough redo of the app. I also improved the
graphical interface, and allowed more than 6 shapes to be displayed.